Update README.md to reflect version of node.js
Because we are using express version 2.5.* (see server/package.json), it
depends on having node.js version 0.8.* installed.

Using the most recent version of node.js (0.10.* at this time) causes an
exception when trying to run the installer.

In the future, the appdev module could be upgraded to express 3.*, which
would allow us to use node.js version 0.10.*