Use app as the testTarget
This way, the test will automatically start its own server and
run the tests against that.

I've added a before() function, which is unimplemented. Right now
the first test is very slow, because the server is not fully started.
If we can wait for this in the before function, then that test's
timing would be reported more accurately.